HC-256: Update PostProcessor to use hysds_commons ElasticSearch library
Also did some refactoring and cleanup
