+++ chimera/postprocess_functions.py
@@ -1,12 +1,34 @@
+import traceback
+import time
+import json
+
+from hysds.es_util import get_grq_es, get_mozart_es
+
 from chimera.commons.accountability import Accountability
+from chimera.commons.constants import ChimeraConstants
+
+from chimera.logger import logger
+
 
 class PostProcessFunctions(object):
-    def __init__(self, context, pge_config, settings, psuedo_context):
+    MOZART_ES_ENDPOINT = "MOZART"
+    GRQ_ES_ENDPOINT = "GRQ"
+
+    def __init__(self, context, pge_config, settings, job_result, mozart_es=None, grq_es=None):
         self._context = context
         self._pge_config = pge_config
         self._settings = settings
-        self._psuedo_context = psuedo_context
+        self._job_result = job_result
         self.accountability = Accountability(self._context)
+        if mozart_es:
+            self._mozart_es = mozart_es
+        else:
+            self._mozart_es = get_mozart_es()
+
+        if grq_es:
+            self._grq_es = grq_es
+        else:
+            self._grq_es = get_grq_es()
 
     def run(self, function_list):
         """
@@ -18,10 +40,10 @@ class PostProcessFunctions(object):
         """
         output_context = dict()
         for func in function_list:
-            self._psuedo_context.update(getattr(self, func)())
+            self._job_result.update(getattr(self, func)())
         
         self.accountability.set_status("job-completed")
-        return self._psuedo_context
+        return self._job_result
 
     def _check_job_status(self):
         """
@@ -53,7 +75,7 @@ class PostProcessFunctions(object):
         :return: tuple(products_staged, prev_context, message)
         """
         job_id = str(self._job_result["payload_id"])
-        endpoint = MOZART_ES_ENDPOINT
+        endpoint = self.MOZART_ES_ENDPOINT
         return_job_id = None
 
         """
@@ -215,6 +237,210 @@ class PostProcessFunctions(object):
                 )
 
         return product_id, products_url_list, products_metadata_list
+
+    def query_es(
+        self,
+        endpoint,
+        doc_id=None,
+        query=None,
+        request_timeout=30,
+        retried=False,
+        size=1,
+    ):
+        """
+        This function queries ES. Not using the query util because the ES
+        connection is set
+        for the GRQ ES.
+        :param endpoint: the value specifies which ES endpoint to send query
+         can be MOZART or GRQ
+        :param doc_id: id of product or job
+        :param query: query to run
+        :param request_timeout: how long to wait for ES request
+        :param retried: flag to specify if the query has already been retried
+        :param size: number of results to be returned
+        :return: result of query
+        """
+        result = None
+        if query is None and doc_id is None:
+            raise ValueError("Both doc_id and query cannot be None")
+
+        es, es_url, es_index = None, None, None
+        if endpoint == self.GRQ_ES_ENDPOINT:
+            es_index = "grq"
+            es = self._grq_es
+        if endpoint == self.MOZART_ES_ENDPOINT:
+            es_index = "job_status-current"
+            es = self._mozart_es
+
+        if doc_id is not None:
+            query = {"query": {"bool": {"must": [{"term": {"_id": doc_id}}]}}}
+
+        try:
+            result = es.search(
+                index=es_index, body=query, size=size, request_timeout=request_timeout
+            )
+            # retry in case of time out
+            if "timed_out" in result and result.get("timed_out"):
+                logger.warning(
+                    "ES responded with a timed out result, "
+                    "retrying....: {}".format(json.dumps(result))
+                )
+                raise RuntimeWarning(
+                    "ES responded with a timed out result, retrying...."
+                )
+        except Exception as e:
+            logger.warning(
+                "Caught exception from elasticsearch "
+                "retrying: {}".format(traceback.format_exc())
+            )
+            # Retry querying, this is incase ES takes too long to respond
+            if not retried:
+                self.query_es(
+                    endpoint=endpoint,
+                    doc_id=doc_id,
+                    size=size,
+                    request_timeout=int(request_timeout + 30),
+                    retried=True,
+                )
+            else:
+                raise Exception(str(e))
+
+        return result
+
+    def product_in_grq(self, doc_id):
+        """
+        Checks if the product has been indexed in ES
+        :param doc_id:
+        :return: True if product found else throw suitable exception
+        """
+        query = {
+            "_source": ["id"],
+            "query": {"bool": {"must": [{"term": {"_id": doc_id}}]}},
+        }
+
+        try:
+            if self.wait_for_doc(endpoint=self.GRQ_ES_ENDPOINT, query=query, timeout=120):
+                return True
+        except Exception as ex:
+            logger.error(
+                "Error querying GRQ for product {}. {}. {}".format(
+                    doc_id, str(ex), traceback.format_exc()
+                )
+            )
+            raise Exception(
+                "Error querying GRQ for product {}. {}".format(doc_id, str(ex))
+            )
+
+    def wait_condition(self, endpoint, result):
+        results_exist = len(result.get("hits").get("hits")) == 0
+        if endpoint == self.MOZART_ES_ENDPOINT:
+            return results_exist or str(result.get("hits").get("hits")[0].get(
+                "_source").get("status")) == "job-started"
+        if endpoint == self.GRQ_ES_ENDPOINT:
+            return results_exist
+
+    def wait_for_doc(self, endpoint, query, timeout):
+        """
+        This function executes the search query for specified wait time until
+        document is found
+        :param endpoint: GRQ or MOZART
+        :param query: search query
+        :param timeout: time to wait in seconds
+        :return: True if document found else raise suitable Exception
+        """
+        try:
+            result = self.query_es(
+                endpoint=endpoint, query=query, request_timeout=30, size=1
+            )
+            slept_seconds = 0
+            sleep_seconds = 2
+
+            while self.wait_condition(endpoint=endpoint, result=result):
+                if result.get("timed_out", True):
+                    slept_seconds += 30
+
+                if slept_seconds + sleep_seconds < timeout:
+                    logger.debug("Slept for {} seconds".format(slept_seconds))
+                    logger.debug("Sleeping for {} seconds".format(sleep_seconds))
+                else:
+                    sleep_seconds = timeout - slept_seconds
+                    logger.debug("Slept for {} seconds".format(slept_seconds))
+                    logger.debug(
+                        "Sleeping for {} seconds to conform to timeout "
+                        "of {} seconds".format(sleep_seconds, timeout)
+                    )
+
+                if slept_seconds >= timeout:
+                    if len(result.get("hits").get("hits")) == 0:
+                        raise Exception(
+                            "{} ES taking too long to index document".format(endpoint)
+                        )
+                    if endpoint == self.MOZART_ES_ENDPOINT:
+                        if (
+                            str(result["hits"]["hits"][0]["_source"]["status"])
+                            == "job-started"
+                        ):
+                            raise Exception(
+                                "{} ES taking too long to update status of "
+                                "job".format(endpoint)
+                            )
+
+                time.sleep(sleep_seconds)
+                result = self.query_es(
+                    endpoint=endpoint, query=query, request_timeout=30, size=1
+                )
+                slept_seconds += sleep_seconds
+                sleep_seconds *= 2
+            return True
+        except Exception as e:
+            raise Exception("ElasticSearch Operation failed due to : {}".format(str(e)))
+
+    def get_product_info(self, product_id):
+        """
+            This function gets the product's URL and associated metadata from Elastic
+            Search
+            :param product_id: id of product
+            :return: tuple(product_url, metadata)
+            """
+        response = None
+        try:
+            if self.product_in_grq(doc_id=product_id):
+                try:
+                    response = self.query_es(
+                        endpoint=self.GRQ_ES_ENDPOINT, doc_id=product_id
+                    )
+                    if len(response.get("hits").get("hits")) == 0:
+                        raise Exception(
+                            "ES taking too long to index product with id "
+                            "%s." % product_id
+                        )
+                except Exception as ex:
+                    raise Exception(
+                        "ElasticSearch Operation failed due to : {}".format(str(ex))
+                    )
+        except Exception as ex:
+            raise Exception(
+                "Failed to find product in GRQ. {}. {}".format(
+                    str(ex), traceback.format_exc()
+                )
+            )
+
+        try:
+            result = response.get("hits").get("hits")[0]
+            product_urls = result.get("_source").get("urls")
+            product_url = None
+            for url in product_urls:
+                if url.startswith("s3://"):
+                    product_url = url
+            metadata = result.get("_source").get("metadata")
+        except Exception as ex:
+            raise Exception(
+                "Failed to get product info. {}. {}".format(
+                    str(ex), traceback.format_exc()
+                )
+            )
+
+        return product_url, metadata
 
     def core_post_process_steps(self):
         """
